+++ CHANGES.rst
@@ -111,6 +111,9 @@ Bug fixes:
 
 - Fixed several exponential backtracking regex issues (`#1435`_).
 
+- Fixed `inspect` command not recognizing MySQL BLOB types (`#1484`_).
+
+  .. _#1484: https://github.com/atviriduomenys/spinta/issues/1484
   .. _#1435: https://github.com/atviriduomenys/spinta/issues/1435
   .. _#1409: https://github.com/atviriduomenys/spinta/issues/1409
   .. _#1269: https://github.com/atviriduomenys/spinta/issues/1269